Added script; Lifted number inits, Library Code
I filled out a few more of the library types for llvm, although I'm
still not entire sure that I am doing it correctly (since I don't
fully understand the OCaml versions). However, I am doing it in such
a way as makes sense given what we are building, so it probably won't
matter much.
I've also added code to lift number initializations to the top of the
main function, and to replace number literals with the relevent LLVM
identifiers, as well as creating the script "pyretc", which handles
both steps of the compilation process (translating the program into
LLVM, and then compiling that with clang).