Make everything use same Bind type.
Moving everything to new Bind types that are mostly the same is silly.
Everything now uses the same bind type, which has two variants: one with
location and one without. There is room for improvement and room for
preserving types better throughout different steps but this will do for
now.