Finish up most Lower->LLVM transformations.
Most transformations to LLVM have been implemented now (but they are
subject to change). As a result, the LLVM representation has
received several improvements and bug fixes.

The identifier names in anf-to-h.arr have had their prefixes stripped.
The binding situation will need to be improved in the compiler, so that
local and global identifiers are chosen correctly.

compile.arr will now produce a *.ll file for some tests (for example
tests/lambdas/simple.arr), but they will not yet compile.