Preserve more types; prepare for closures; fix bugs.
There were some places where we were discarding type annotations
unnecessarily. Since we will eventually want to act on some of this, try
to save it where we can. We now make use of the new types library, which
allows us to specify some types that are not available to the user.

Change how we handle closures. Since we will want to use trampolines to
implement closures, we don't want to modify the arguments or call
structure of the functions. This will allow us to pass around a function
pointer instead of a closure record.

Fix some minor bugs, including logic for the branches we generate for
cases statements.